...G...PS. [ZBXNEXT-101] fixed function argument type
git-svn-id: svn://svn.zabbix.com/branches/dev@49140 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-101/src/libs/zbxnix/control.c
@@ -19,7 +19,7 @@
 
 #include "control.h"
 
-static int	parse_log_level_options(const char *opt, int len, int *scope, int *data)
+static int	parse_log_level_options(const char *opt, size_t len, int *scope, int *data)
 {
 	int		num = 0;
 	const char	*rtc_options;
